2024/Codegate/Chatting-Service/Flask/src/requirements.txt
aioflask==0.4.0
asgiref==3.7.2
blinker==1.7.0
click==8.1.7
configparser==6.0.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
greenletio==0.11.0
h11==0.14.0
hashring==1.5.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
memcache==0.12.0
mysql-connector-python==8.3.0
mysqlclient==2.2.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
pymemcache==4.0.0
PyMySQL==1.1.0
SQLAlchemy==2.0.28
typing_extensions==4.10.0
uvicorn==0.28.0
Werkzeug==3.0.1